+++ bindings/pydeck/docs/CHANGELOG.rst
@@ -8,7 +8,7 @@ Releases and associated GitHub PRs for pydeck are documented here.
 
 0.8.0b3 - Sep 09 2022
 ^^^^^^^^^^^^^^^^^^^^^
-- Add configuration. Fix css_background_color (#7249)
+- Add configuration attribute in pydeck. Fix css_background_color (#7255)
 
 0.8.0b2 - Sep 08 2022
 ^^^^^^^^^^^^^^^^^^^^^